misc: lint solidity code

+++ contracts/Interfaces/Compound/CTokenWithFlashloan.sol
@@ -1,6 +1,10 @@
 import "./CErc20I.sol";
 
-interface CTokenWithFlashloan  is CErc20I {
+interface CTokenWithFlashloan is CErc20I {
     //Cream's custom flashloan implementation
-    function flashLoan(address receiver, uint amount, bytes calldata params) external;
+    function flashLoan(
+        address receiver,
+        uint256 amount,
+        bytes calldata params
+    ) external;
 }
